Split visual model loading into platform specific files
